Don't initialize `uinput` unless it's necessary
I'm only using `cecdaemon` to run commands, not to send keypresses via
`uinput`. I'd like to avoid initializing `uinput` because it's a bit
tricky to set up:

1. It requires the user I'm running `cecdaemon` under to permissions it
   doesn't currently have.
2. It requires my system to have loaded the `uinput` kernel module,
   which doesn't actually happen on my device. See
   https://github.com/pyinput/python-uinput/issues/9#issuecomment-3568271848
   for more details.

While I could fix both of these issues, this is such a minor change to
`cecdaemon`, I thought it would be an acceptable change.